# Change `bn256Add` to use consensys gnark-bn254 lib