--- a/core/vm/contracts.go
+++ b/core/vm/contracts.go
@@ -31,8 +31,6 @@ import (
 	"github.com/consensys/gnark-crypto/ecc/bls12-381/fr"
 	"github.com/holiman/uint256"
 
-	evmone "github.com/erigontech/evmone_precompiles"
-
 	"github.com/erigontech/erigon-lib/chain"
 	"github.com/erigontech/erigon-lib/chain/params"
 	"github.com/erigontech/erigon-lib/common"
@@ -40,6 +38,7 @@ import (
 	"github.com/erigontech/erigon-lib/crypto"
 	"github.com/erigontech/erigon-lib/crypto/blake2b"
 	"github.com/erigontech/erigon-lib/crypto/bn256"
+	"github.com/consensys/gnark-crypto/ecc/bn254"
 	libkzg "github.com/erigontech/erigon-lib/crypto/kzg"
 	"github.com/erigontech/erigon-lib/crypto/secp256r1"
 	"github.com/erigontech/erigon/core/tracing"
@@ -605,14 +604,19 @@ func newTwistPoint(blob []byte) (*bn256.G2, error) {
 // runBn256Add implements the Bn256Add precompile, referenced by both
 // Byzantium and Istanbul operations.
 func runBn256Add(input []byte) ([]byte, error) {
-	x := getData(input, 0, 64)
-	y := getData(input, 64, 64)
-	var out [64]byte
-	if evmone.EcAdd(&out, (*[64]byte)(x), (*[64]byte)(y)) {
-		return out[:], nil
-	} else {
-		return nil, errors.New("bn256: invalid point")
+	x := &bn254.G1Affine{}
+	_, err := x.SetBytes(getData(input, 0, 64))
+	if err != nil {
+		return nil, err
 	}
+
+	y := &bn254.G1Affine{}
+	_, err = y.SetBytes(getData(input, 64, 64))
+	if err != nil {
+		return nil, err
+	}
+	x = x.Add(x, y)
+	return x.Marshal(), nil
 }
 
 // bn256Add implements a native elliptic curve point addition conforming to